Modified behavior to treat "floating" events found in a calendar with X-WR-TIMEZONE as if they are in the time zone specified
The code was crashing because of the check in is_UTC not checking if a
timezone existed or not.  If there was no timezone, then the code
crashed.  This matches the case of a "floating" event, which is normally
interpreted as being in the current time zone.

This change treats these "floating" events as if they are in the
timezone specified by X-WR-TIMEZONE, and not as "floating" events.
Based on https://github.com/u01jmg3/ics-parser/issues/245 and
https://blog.jonudell.net/2011/10/17/x-wr-timezone-considered-harmful/,
and a calendar a user sent me, I think this is what was intended by the
calendars' authors.  It's hard to say for sure, though, since I've been
unable to find any definitive documentation, only educated guesses.